requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
oauthlib==3.2.2
osrparse==6.0.2
ossapi==4.0.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
typing-utils==0.1.0
urllib3==2.2.2
yarl==1.9.4
